+++ tests/testthat/httpbin/httpbin.org/status/500.R
@@ -1,0 +1,8 @@
+structure(list(method = "GET", url = "https://httpbin.org/status/500", 
+    status_code = 500L, headers = structure(list(date = "Sun, 07 Dec 2025 18:51:31 GMT", 
+        `content-type` = "text/html; charset=utf-8", `content-length` = "0", 
+        server = "gunicorn/19.9.0", `access-control-allow-origin` = "*", 
+        `access-control-allow-credentials` = "true"), class = "httr2_headers"), 
+    body = raw(0), timing = c(redirect = 0, namelookup = 0, connect = 0, 
+    pretransfer = 1e-04, starttransfer = 0.278247, total = 0.27831
+    ), cache = new.env(parent = emptyenv())), class = "httr2_response")